feat: connector 를 이용하여 test db와 product db 분리
