+++ src/main/java/repository/JdbcConnector.java
@@ -1,0 +1,7 @@
+package repository;
+
+import java.sql.Connection;
+
+public interface JdbcConnector {
+    Connection getConnection();
+}
